+++ lib/voicemail/plugin.rb
@@ -33,7 +33,8 @@ module Voicemail
         pin_wrong "The PIN you entered does not match. Please try again.", desc: "Message for an user that enters the wrong PIN"
         could_not_auth "We are sorry, the system could not authenticate you", desc: "Message for authentication final failure."
         number_before "You have ", desc: "What to play before the number of new messages"
-        number_after " new messages", desc: "What to play after the number of new messages"
+        number_after_new " new messages", desc: "What to play after the number of new messages"
+        number_after_saved " saved messages", desc: "What to play after the number of saved messages"
         menu_greeting "Press 1 to listen to new messages, 2 to change your greeting, 3 to change your PIN", desc: "What to say before the main menu"
         menu_timeout_message "Please enter a digit for the menu", desc: "Message to play on main menu timeout"
         menu_invalid_message "Please enter valid input", desc: "Message to play on main menu invalid"
@@ -67,6 +68,7 @@ module Voicemail
       messages {
         menu "Press 1 to archive the message and go to the next, press 5 to delete the message and go to the next, press 7 to hear the message again, press 9 for the main menu", desc: "Menu to use inside messages"
         no_new_messages "There are no new messages", desc: "Message to inform the user he has no new messages"
+        no_saved_messages "There are no saved messages", desc: "Message to inform the user he has no saved messages"
         message_received_on "Message received on ", desc: "Prefix to menu intro"
         from " from ", desc: "Used in message intro"
       }